handheld_device_client/requirements.txt
hidapi==0.11.0.post2
requests==2.27.1
